Finished version of projective transform. The main bug was dealing with the fact that matrices are indexed by (row, col) but all the math for figuring out bounding box and inverse-mapped points is done with (x, y), where (x -> col, y -> row).
There was also a problem with leaving a byte in the stream before
reading the rest of a binary image in both Pgm and Ppm. In Ppm this was
more apparent because all Vector3b values would be shifted by one, but
in a Pgm image this was not as obvious.

+++ include/matrix/Matrix.hpp
@@ -63,13 +63,13 @@ public:
 
     char* bytes(void) { return reinterpret_cast<char*>(data_.get()); }
 
-    size_t size(void) const { return nelements_; }
+    int32_t size(void) const { return nelements_; }
 
-    size_t size_in_bytes(void) const { return nbytes_; }
+    int32_t size_in_bytes(void) const { return nbytes_; }
 
 private:
-    size_t nelements_;
-    size_t nbytes_;
+    int32_t nelements_;
+    int32_t nbytes_;
     std::unique_ptr<Dtype[]> data_;
 };
 
@@ -106,6 +106,8 @@ Matrix<Dtype, R1, C2> operator*(const Matrix<Dtype, R1, Dynamic>& m1,
     }
     return mat;
 }
+
+/*
 
 // Partial specialization for Vector types
 template <typename Dtype, int32_t Length>
@@ -171,6 +173,7 @@ private:
     size_t nbytes_;
     std::unique_ptr<Dtype[]> data_;
 };
+*/
 }
 
 #endif
